tabix/index/reference_sequence/bin: Split max ID and metadata ID

+++ noodles-tabix/src/index/reference_sequence.rs
@@ -208,7 +208,7 @@ impl ReferenceSequence {
 fn region_to_bins(start: usize, mut end: usize) -> BitVec {
     end -= 1;
 
-    let mut bins = BitVec::from_elem(bin::MAX_ID, false);
+    let mut bins = BitVec::from_elem(bin::MAX_ID as usize, false);
     bins.set(0, true);
 
     for k in (1 + (start >> 26))..=(1 + (end >> 26)) {
@@ -257,7 +257,7 @@ mod tests {
     fn test_region_to_bins() {
         // [8, 13]
         let actual = region_to_bins(7, 13);
-        let mut expected = BitVec::from_elem(bin::MAX_ID, false);
+        let mut expected = BitVec::from_elem(bin::MAX_ID as usize, false);
         for &k in &[0, 1, 9, 73, 585, 4681] {
             expected.set(k, true);
         }
@@ -265,7 +265,7 @@ mod tests {
 
         // [63245986, 63245986]
         let actual = region_to_bins(63245985, 63255986);
-        let mut expected = BitVec::from_elem(bin::MAX_ID, false);
+        let mut expected = BitVec::from_elem(bin::MAX_ID as usize, false);
         for &k in &[0, 1, 16, 133, 1067, 8541] {
             expected.set(k, true);
         }
